pyproject.toml
[tool.poetry]
name = "github-events-api"
version = "0.1.0"
description = ""
authors = ["Iva Kocourkova <iva@kocourkovi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
sqlmodel = "^0.0.21"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pre-commit = "^3.8.0"
mypy = "^1.11.1"
ruff = "^0.6.2"
types-requests = "^2.32.0.20240712"
types-pyyaml = "^6.0.12.20240808"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]

[tool.mypy]
ignore_missing_imports = true
python_version = "3.11"
files = "^(github_events_api)"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
